+++ app/models/purchase_address.rb
@@ -16,7 +16,7 @@ class PurchaseAddress
 
   def save
     purchase = Purchase.create(item_id: item_id, user_id: user_id)
-    Address.create(post_code: post_code, prefecture_id: prefecture_id, city: city, street: street, building_name: building_name, phone_number: phone_number, purchase_id: purchase_id)
+    Address.create(purchase_id: purchase.id, post_code: post_code, prefecture_id: prefecture_id, city: city, street: street, building_name: building_name, phone_number: phone_number)
   end
 
 end
