+++ app/views/items/index.html.erb
@@ -159,7 +159,7 @@
 
       <% if @items.empty? %>
       <li class='list'>
-        <%= link_to item_path(item) do %>
+        <%= link_to item_path(@items) do %>
         <%= image_tag "https://tech-master.s3.amazonaws.com/uploads/curriculums/images/Rails1-4/sample.jpg", class: "item-img" %>
         <div class='item-info'>
           <h3 class='item-name'>
