+++ app/views/items/show.html.erb
@@ -29,7 +29,7 @@
     <%= link_to "削除", item_path{item.id} , data: {turbo_method: :delete}, class:"item-destroy" %>
 
     <% elsif %>
-    <%= link_to "購入画面に進む", "#", data: { turbo: false }, class:"item-red-btn"%>
+    <%= link_to "購入画面に進む", item_purchases_path(item_id: @item.id) , data: { turbo: false }, class:"item-red-btn"%>
 
     <% end %>
    <% end %>
