+++ config/routes.rb
@@ -3,6 +3,6 @@ Rails.application.routes.draw do
   root to: 'items#index'
 
   resources :items do
-    references :purchases only: [:index, :create]
+    resources :purchases, only: [:index, :create]
   end
 end
