+++ db/migrate/20240519054836_create_purchases.rb
@@ -1,0 +1,9 @@
+class CreatePurchases < ActiveRecord::Migration[7.0]
+  def change
+    create_table :purchases do |t|
+  t.references :user, null: false, foreign_key: true
+  t.references :item, null: false, foreign_key: true
+  t.timestamps
+    end
+  end
+end
