--- db/migrate/20240520035039_create_purchases.rb
@@ -1,9 +1,0 @@
-class CreatePurchases < ActiveRecord::Migration[7.0]
-  def change
-    create_table :purchases do |t|
-      t.references :user, null: false, foreign_key: true
-      t.references :item, null: false, foreign_key: true
-      t.timestamps
-    end
-  end
-end
